requirements.txt
pandas
cassio
torch
clip
pillow
langchain
